// src/types/resources/Consent.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TReference } from '../partials/Reference';
import { TDateTime } from '../simpleTypes/DateTime';
import { TAttachment } from '../partials/Attachment';
import { TConsentPolicy } from '../partials/ConsentPolicy';
import { TConsentVerification } from '../partials/ConsentVerification';
import { TConsentProvision } from '../partials/ConsentProvision';

export type TConsent = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  status: string;
  scope: TCodeableConcept;
  category: TCodeableConcept[];
  patient?: TReference;
  dateTime?: TDateTime;
  performer?: TReference[];
  organization?: TReference[];
  sourceAttachment?: TAttachment;
  sourceReference?: TReference;
  policy?: TConsentPolicy[];
  policyRule?: TCodeableConcept;
  verification?: TConsentVerification[];
  provision?: TConsentProvision;
};